services/files-svc/requirements.txt
fastapi==0.115.0
uvicorn[standard]==0.30.0
python-multipart==0.0.9
pydantic==2.8.2
pydantic-settings==2.4.0
python-jose==3.3.0
minio==7.2.9
